+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Scraping of tenant cluster prometheus
 - Scraping of control plane prometheus
+- Add installation label
+
+### Changed
+
+- Set honor labels to true
 
 ## [1.2.0] - 2020-09-03